Moved protocol conformance to unique files
