+++ Source/CommonMarkAST.swift
@@ -8,45 +8,15 @@
 
 import Foundation
 
-// MARK: Protocols
-
-internal protocol CommonMarkRenderable {
-  var commonMark: String { get }
-}
-
-internal protocol HTMLRenderable {
-  var html: String { get }
-}
-
 // MARK: Enums
 
 internal enum HeadingLevel: Int, CaseIterable {
   case h1 = 1, h2, h3, h4, h5, h6
 }
 
-internal enum CommonMarkAST: Equatable, CommonMarkRenderable, HTMLRenderable {
+internal enum CommonMarkAST: Equatable {
   case heading(level: HeadingLevel, nodes: [CommonMarkAST])
   case text(_ text: String)
-
-  // MARK: CommonMarkRenderable
-  var commonMark: String {
-    switch self {
-    case let .heading(level, nodes):
-      return String(repeating: "#", count: level.rawValue) + " " + nodes.commonMark + "\n\n"
-    case let .text(text):
-      return text.sanatizeHTML()
-    }
-  }
-
-  // MARK: HTMLRenderable
-  var html: String {
-    switch self {
-    case let .heading(level, nodes):
-      return "<\(level)>" + nodes.html + "</\(level)>\n"
-    case let .text(text):
-      return text
-    }
-  }
 
   // MARK: Node Type
   var type: NodeType {
@@ -97,35 +67,5 @@ internal enum CommonMarkAST: Equatable, CommonMarkRenderable, HTMLRenderable {
       level: HeadingLevel(rawValue: captures.first?.count ?? 0) ?? .h1,
       nodes: markdown.isEmpty ? [] : [.text(markdown)]
     )
-  }
-
-  init(text: String) {
-    self = .text(text)
-  }
-}
-
-// MARK: Extensions
-
-extension Array where Element == CommonMarkAST {
-  var html: String { return map { $0.html }.joined() }
-  var commonMark: String { return map { $0.commonMark }.joined() }
-}
-
-// MARK: - Extensions
-
-private extension String {
-  func sanatizeHTML() -> String {
-    return map { $0.sanatizeHTML() }.joined()
-  }
-}
-
-private extension Character {
-  func sanatizeHTML() -> String {
-    switch self {
-    case "<": return "&lt;"
-    case ">": return "&gt;"
-    case "\"": return "&quot;"
-    default: return String(self)
-    }
   }
 }
